requirements.txt
langchain
langchain-community
langchain-ollama
openai
python-dotenv
geopy
folium
pandas
